+++ src/Plugins/SmartStore.AmazonPay/Controllers/AmazonPayShoppingCartController.cs
@@ -55,19 +55,18 @@ namespace SmartStore.AmazonPay.Controllers
 			return new EmptyResult();
 		}
 
-		[ChildActionOnly]
-		public ActionResult WidgetLibrary()
-		{
-			// Not possible to load it asynchronously cause of document.write inside.
-			//var widgetUrl = _apiService.GetWidgetUrl();
+		//[ChildActionOnly]
+		//public ActionResult WidgetLibrary()
+		//{
+		//	// Not possible to load it asynchronously cause of document.write inside.
+		//	var widgetUrl = _apiService.GetWidgetUrl();
 
-			//if (widgetUrl.HasValue())
-			//{
-			//	//return this.Content("<script src=\"{0}\" type=\"text/javascript\"></script>".FormatWith(widgetUrl));
-			//	return Content($"<script async=\"async\" src=\"{widgetUrl}\" type=\"text/javascript\"></script>");
-			//}
+		//	if (widgetUrl.HasValue())
+		//	{
+		//		return this.Content("<script src=\"{0}\" type=\"text/javascript\"></script>".FormatWith(widgetUrl));
+		//	}
 
-			return new EmptyResult();
-		}
+		//	return new EmptyResult();
+		//}
 	}
 }