+++ src/Plugins/SmartStore.AmazonPay/changelog.md
@@ -1,5 +1,10 @@
 ﻿#Release Notes
 
+##Login and Pay with Amazon 3.0.1
+###Improvements
+* Implemented new Login and Pay with Amazon
+	* Widget script is now loaded asynchronously.
+
 ##Pay with Amazon 2.2.0.3
 ###Improvements
 * Added message token %SmartStore.AmazonPay.BillingAddressMessageNote% for billing address note in order placed customer notification
@@ -30,7 +35,7 @@
 
 ##Pay with Amazon 1.16
 ###Bugfixes
-* Reflect refunds made at amazon seller central when using data polling
+* Reflect refunds made at Amazon Seller Central when using data polling
 
 ##Pay with Amazon 1.15
 ###Bugfixes
@@ -38,7 +43,7 @@
 
 ##Pay with Amazon 1.14
 ###Bugfixes
-* Data polling did not reflect the transaction status correctly if the action took place at amazon seller central
+* Data polling did not reflect the transaction status correctly if the action took place at Amazon Seller Central
 
 ##Pay with Amazon 1.13
 ###Bugfixes